LabWorks (ability to mark completed or not)

+++ AcademicProgressTracker.Domain/Entities/LabWork.cs
@@ -4,7 +4,7 @@
     {
         public Guid Id { get; set; }
         public int Number { get; set; }
-        public float MaximumScore { get; set; }
+        public decimal MaximumScore { get; set; }
 
         public Guid SubjectId { get; set; }
         public Subject? Subject { get; set; }
